refactor: full refactoring of the ShippingLabelPackageDetail screen + ShippingLabelPackageList + view model
