+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/Create Shipping Label Form/ShippingLabelFormViewController.swift
@@ -1,5 +1,6 @@
 import UIKit
 import Yosemite
+import SwiftUI
 
 final class ShippingLabelFormViewController: UIViewController {
