+++ WooCommerce/WooCommerceTests/ViewRelated/Shipping Label/Create Shipping Label/ShippingLabelPackageDetailsViewModelTests.swift
@@ -14,13 +14,17 @@ final class ShippingLabelPackageDetailsViewModelTests: XCTestCase {
         storageManager.viewStorage
     }
 
+    private var stores: MockStoresManager!
+
     override func setUp() {
         super.setUp()
         storageManager = MockStorageManager()
+        stores = MockStoresManager(sessionManager: SessionManager.makeForTesting(authenticated: true))
     }
 
     override func tearDown() {
         storageManager = nil
+        stores = nil
         super.tearDown()
     }
 
@@ -83,8 +87,67 @@ final class ShippingLabelPackageDetailsViewModelTests: XCTestCase {
 
         waitForExpectations(timeout: Constants.expectationTimeout, handler: nil)
     }
+
+    func test_didSelectPackage_returns_the_expected_value() {
+        // Given
+        let customPackage = ShippingLabelCustomPackage(isUserDefined: true,
+                                                       title: "Box",
+                                                       isLetter: true,
+                                                       dimensions: "3 x 10 x 4",
+                                                       boxWeight: 10,
+                                                       maxWeight: 11)
+        let order = MockOrders().empty().copy(siteID: sampleSiteID)
+        let currencyFormatter = CurrencyFormatter(currencySettings: CurrencySettings())
+        mockRetrieveShippingLabelPackageDetails(result: .success(mockPackageResponse()))
+        let viewModel = ShippingLabelPackageDetailsViewModel(order: order,
+                                                             formatter: currencyFormatter,
+                                                             stores: stores,
+                                                             storageManager: storageManager,
+                                                             weightUnit: "kg")
+
+
+        XCTAssertNil(viewModel.selectedCustomPackage)
+        XCTAssertNil(viewModel.selectedPredefinedPackage)
+
+        // When
+        viewModel.didSelectPackage(customPackage.title)
+
+        // Then
+        XCTAssertEqual(viewModel.selectedCustomPackage, customPackage)
+        XCTAssertNil(viewModel.selectedPredefinedPackage)
+    }
+
+    func test_confirmPackageSelection_returns_the_expected_value() {
+        // Given
+        let customPackage = ShippingLabelCustomPackage(isUserDefined: true,
+                                                       title: "Box",
+                                                       isLetter: true,
+                                                       dimensions: "3 x 10 x 4",
+                                                       boxWeight: 10,
+                                                       maxWeight: 11)
+        let order = MockOrders().empty().copy(siteID: sampleSiteID)
+        let currencyFormatter = CurrencyFormatter(currencySettings: CurrencySettings())
+        mockRetrieveShippingLabelPackageDetails(result: .success(mockPackageResponse()))
+        let viewModel = ShippingLabelPackageDetailsViewModel(order: order,
+                                                             formatter: currencyFormatter,
+                                                             stores: stores,
+                                                             storageManager: storageManager,
+                                                             weightUnit: "kg")
+
+        XCTAssertNil(viewModel.selectedCustomPackage)
+        XCTAssertNil(viewModel.selectedPredefinedPackage)
+
+        // When
+        viewModel.didSelectPackage(customPackage.title)
+        viewModel.confirmPackageSelection()
+
+        // Then
+        XCTAssertEqual(viewModel.selectedPackageID, customPackage.title)
+    }
+
 }
 
+// MARK: - Utils
 private extension ShippingLabelPackageDetailsViewModelTests {
     func insert(_ readOnlyOrderProduct: Yosemite.Product) {
         let product = storage.insertNewObject(ofType: StorageProduct.self)
@@ -94,5 +157,56 @@ private extension ShippingLabelPackageDetailsViewModelTests {
     func insert(_ readOnlyOrderProductVariation: Yosemite.ProductVariation) {
         let productVariation = storage.insertNewObject(ofType: StorageProductVariation.self)
         productVariation.update(with: readOnlyOrderProductVariation)
+    }
+
+    func mockRetrieveShippingLabelPackageDetails(result: Result<ShippingLabelPackagesResponse, Error>) {
+        stores.whenReceivingAction(ofType: ShippingLabelAction.self) { action in
+            if case let ShippingLabelAction.packagesDetails(_, completion: completion) = action {
+                completion(result)
+            }
+        }
+    }
+}
+
+// MARK: - Mocks
+private extension ShippingLabelPackageDetailsViewModelTests {
+    func mockPackageResponse() -> ShippingLabelPackagesResponse {
+        let storeOptions = ShippingLabelStoreOptions(currencySymbol: "$",
+                                                     dimensionUnit: "in",
+                                                     weightUnit: "oz",
+                                                     originCountry: "US")
+
+        let customPackages = [
+            ShippingLabelCustomPackage(isUserDefined: true,
+                                       title: "Box",
+                                       isLetter: true,
+                                       dimensions: "3 x 10 x 4",
+                                       boxWeight: 10,
+                                       maxWeight: 11),
+            ShippingLabelCustomPackage(isUserDefined: true,
+                                       title: "Box n°2",
+                                       isLetter: true,
+                                       dimensions: "30 x 1 x 20",
+                                       boxWeight: 2,
+                                       maxWeight: 4),
+            ShippingLabelCustomPackage(isUserDefined: true,
+                                       title: "Box n°3",
+                                       isLetter: true,
+                                       dimensions: "10 x 40 x 3",
+                                       boxWeight: 7,
+                                       maxWeight: 10)]
+
+        let predefinedOptions = [ShippingLabelPredefinedOption(title: "USPS", predefinedPackages: [ShippingLabelPredefinedPackage(id: "package-1",
+                                                                                                                                  title: "Small",
+                                                                                                                                  isLetter: true,
+                                                                                                                                  dimensions: "3 x 4 x 5"),
+                                                                                                   ShippingLabelPredefinedPackage(id: "package-2",
+                                                                                                                                  title: "Big",
+                                                                                                                                  isLetter: true,
+                                                                                                                                  dimensions: "5 x 7 x 9")])]
+
+        let packagesResponse = ShippingLabelPackagesResponse(storeOptions: storeOptions, customPackages: customPackages, predefinedOptions: predefinedOptions)
+
+        return packagesResponse
     }
 }
